+++ lib/components/Button/Button.tsx
@@ -1,11 +1,11 @@
-import type { LinkProps } from 'next/link'
+import { default as NextLink, type LinkProps } from 'next/link'
 import { forwardRef, PropsWithChildren, ReactNode, Ref } from 'react'
 import { AriaButtonProps } from 'react-aria'
 import { Button as RACButton, ButtonProps as RACButtonProps } from 'react-aria-components'
 
 import cn from '../../tools/cn'
-import { useButtonContext } from './ButtonContextProvider'
 import { ArrowDownIcon, ArrowRightIcon, ExportIcon } from '../../assets/icons'
+import { useComponentLibraryContext } from '../providers/ComponentLibraryProvider'
 import Spinner from '../Spinner/Spinner'
 
 /**
@@ -97,8 +97,8 @@ const Button = forwardRef<HTMLAnchorElement | HTMLButtonElement, ButtonProps>(
     },
     ref,
   ) => {
-    const { LinkComponent } = useButtonContext()
-    const Link = LinkComponent || 'a'
+    const { LinkComponent } = useComponentLibraryContext()
+    const Link = LinkComponent || NextLink
 
     const isLoadingOrDisabled = isLoading || isDisabled
